refactor: RoutineStep 디렉토리 대신 파일로 축소
