--- src/RoomNew/steps/RoutineStep/styles.ts
@@ -1,9 +1,0 @@
-// 플러스(+), 마이너스(-) 버튼에 적용할 스타일
-export const iconButtonStyle =
-  'cursor-pointer text-light-point transition-all hover:text-light-point-hover';
-
-// 루틴 스텝의 각 섹션에서 헤딩에 적용할 스타일
-export const headingStyle = 'mb-6';
-
-// 루틴 스텝의 각 섹션에 적용할 스타일
-export const detailSectionStyle = 'mb-8 text-xl leading-[2rem]';
